HW_14/contact_api_old/pyproject.toml
[tool.poetry]
name = "contact-api"
version = "0.1.0"
description = ""
authors = ["Olena Poroskun <alenka200124@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.28.0"}
sqlalchemy = "^2.0.28"
psycopg2 = "^2.9.9"
python-multipart = "^0.0.9"
crud = "^0.1"
pydantic = {extras = ["email"], version = "^2.6.4"}
alembic = "^1.13.1"
libgravatar = "^1.0.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
fastapi-mail = "^1.4.1"
pytz = "^2024.1"
fastapi-limiter = "^0.1.6"
python-dotenv = "^1.0.1"
cloudinary = "^1.39.1"
pydantic-settings = "^2.2.1"


[tool.poetry.group.dev.dependencies]
sphinx = "^7.2.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
